Refactor docker run into own module. Refactor docker run test into a integration test.

[ci skip-rust]

[ci skip-build-wheels]
