+++ src/python/pants/backend/docker/rules.py
@@ -5,6 +5,7 @@ from pants.backend.docker.dependencies import rules as dependencies_rules
 from pants.backend.docker.docker_binary import rules as binary_rules
 from pants.backend.docker.docker_build import rules as build_rules
 from pants.backend.docker.docker_build_context import rules as context_rules
+from pants.backend.docker.docker_run import rules as run_rules
 from pants.backend.docker.dockerfile_parser import rules as parser_rules
 from pants.backend.docker.publish import rules as publish_rules
 from pants.backend.docker.subsystem import rules as subsystem_rules
@@ -18,5 +19,6 @@ def rules():
         *dependencies_rules(),
         *parser_rules(),
         *publish_rules(),
+        *run_rules(),
         *subsystem_rules(),
     ]
